+++ zapzap/assets/icons/system_icon.py
@@ -46,6 +46,13 @@ class SystemIcon:
             </svg>
             """
         ),
+        "whatsapp_lock": (
+            """
+            <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" fill="{fill_color}" viewBox="0 0 256 256">
+                <path d="M208,80H176V56a48,48,0,0,0-96,0V80H48a16,16,0,0,0-16,16V216a16,16,0,0,0,16,16H208a16,16,0,0,0,16-16V96A16,16,0,0,0,208,80ZM96,56a32,32,0,0,1,64,0V80H96ZM208,216H48V96H208V216Zm-68-60a12,12,0,1,1-20.48-8.48A12,12,0,0,1,140,156Z"/>
+            </svg>
+            """
+        ),
         "flatpak_help": (
             """
             <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" fill="{fill_color}" viewBox="0 0 256 256">